fix: remove ATS settings, add NSAllowsArbitraryLoads
As of iOS 17, HTTPS requests to arbitrary IP addresses are blocked by
default. We could add the IP addresses to the ATS settings explicitly,
but this would not work with the next IP rotation, since it's not
possible to set ATS domains dynamically. The only solution in this case
is to allow arbitrary loads by the client and rely on TrustKit for
verifying trusted certificates.

Jira-Id: VPNAPPL-2249