fix: prevent crash when setting feature flags
Modification of feature flags from ProtonCoreFeatureFlags happens in a
block dispatched synchronously to the main thread. Since TCA 1.14.0,
there is another DispatchQueue.main.sync block that executes in response
to user defaults changed. This causes a deadlock. We need to
temporarily downgrade TCA to 1.13.1 before this sync() call was
introduced