fix(ci): make UI tests run in scientist envs
After spending way too much time to figure this out, it turns out that
the ProtonVPNUITests base case was clobbering the arguments passed by
the scheme. Adding a simple check to forward the arguments (and making
sure `ATLAS_SECRET` is set in the pipeline env) should finally address
this issue.

Jira-Id: VPNAPPL-2127